Renaming ssd1306_i2c to SSD1306
